% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats.R
\name{cats}
\alias{cats}
\title{Expresses your opinions about cats}
\usage{
cats(love = TRUE)
}
\arguments{
\item{love}{A logical argument indicating whether or not you love cats (default = `TRUE`)}
}
\value{
One of two possible character strings (`"I love cats!"` or `"I haven't met a cat yet."`).
}
\description{
Expresses your opinions about cats
}
\examples{
cats(TRUE)
}
