% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{owners}
\alias{owners}
\title{A tibble of example pet owners.}
\format{A tibble with 5 rows and 2 variables:
\describe{
  \item{owner}{the name of a pet owner}
  \item{pet}{the name of a pet}
}}
\usage{
owners
}
\description{
A tibble of example pet owners.
}
\keyword{datasets}
