% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_function.R
\name{pet_function}
\alias{pes_function}
\title{A Function for Prospective Pet Owners}
\usage{
pets_function(x)
}
\arguments{
\item{x}{an integer}
}
\value{
Message on whether user should get a cat or dog
}
\description{
This function, based on the very complicated space-time continum properties of a random number generator to determine whether the user should get a cat or a dog.
}
\examples{
pet_function(2)
pet_function(827)
}
