% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pets.R
\docType{package}
\name{pets}
\alias{pets}
\title{Pets: A package for exclaiming your love of pets.}
\description{
The Pets package provides an incredibly important tool
to voice your love of doggies.
}
\section{Pets functions}{

Pets has one function: say_msg()
}

