% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_msg.R
\name{say_msg}
\alias{say_msg}
\title{A Happy Pet Message}
\usage{
say_msg(my_fave = "puppies", enthusiasm = 3)
}
\arguments{
\item{my_fave}{Your favorite animal. Default = "puppies".}
}
\description{
This function expresses your love of pets.
}
\examples{
say_msg("little dogs")
say_msg()
}
\keyword{pets}
