% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats.R
\name{cats}
\alias{cats}
\title{Expresses your opinion about cats}
\usage{
cats(love = TRUE)
}
\arguments{
\item{love}{A logical argument indicating whether you like cats}
}
\value{
One of two possible char strings, depending on feeligns
}
\description{
Expresses your opinion about cats
}
\examples{
cats(TRUE)
}
