% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats.R
\name{cats}
\alias{cats}
\title{Expresses your feelings about cats}
\usage{
cats(love = TRUE)
}
\arguments{
\item{love}{A logical indicator indicating if you love cats}
}
\value{
one or two character strings (\verb{"I love cats} or \verb{"I don't love cats})
}
\description{
Expresses your feelings about cats
}
\examples{
cats(TRUE)
}
