% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats.R
\name{cats}
\alias{cats}
\title{Expresses your feelings about pets}
\usage{
cats(love = TRUE)
}
\arguments{
\item{love}{A logical argument indicating whether or not you like cats}
}
\value{
One of two character messages about cat feelings e.g. \verb{I love cats}
}
\description{
Expresses your feelings about pets
}
\examples{
cats(TRUE)
}
