% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats.R
\name{cats}
\alias{cats}
\title{Express your opinions about cats}
\usage{
cats(love = TRUE)
}
\arguments{
\item{love}{A logical argument indicating whether or not you love cats (default = \code{TRUE})}
}
\value{
One of two possible character strings ("I love cats!" or "I am not a cat person.").
}
\description{
Express your opinions about cats
}
\examples{
cats(TRUE)
}
