% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats.R
\name{cats}
\alias{cats}
\title{Expresses feelings about cats}
\usage{
cats(love = TRUE)
}
\arguments{
\item{love}{A logical argument indicating if you love cats. Either TRUE or FALSE}
}
\value{
returns one of two possible character strings describing your feelings about cats
}
\description{
Expresses feelings about cats
}
\examples{
cats(TRUE)
}
