% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats.R
\name{cats}
\alias{cats}
\title{Expressed your feelings about cats}
\usage{
cats(love = TRUE)
}
\arguments{
\item{love}{A logical argument indicating whether or not you like cats}
}
\value{
One of two character messages about cat feelings
}
\description{
Expressed your feelings about cats
}
\examples{
cats(TRUE)
}
