% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pets}
\alias{pets}
\title{A tibble of example pets.}
\format{A tibble with 5 rows and 2 variables:
\describe{
  \item{pet}{the name of a pet}
  \item{species}{the species of the pet}
}}
\usage{
pets
}
\description{
A tibble of example pets.
}
\keyword{datasets}
